🧪 [testing] add missing sysctl error tests and fix pointer truncation (#1264)
Adds a test to `tests/lib/c/test_sysctl.c` to verify that `sysctl` correctly
handles and translates the return value from the underlying syscall when it
fails (returning -1 and setting errno appropriately).

During test implementation, an issue was discovered and fixed in
`lib/c/src/sysctl_helpers.c`: arguments passed to `_syscall6` were
being cast to `(int)`, which truncates pointers on 64-bit platforms
and causes segmentation faults. The definition of `_syscall6` and
the calls were updated to safely use `intptr_t`.

Additionally, the test file was refactored to `#include` the actual
source file rather than relying on inline duplicates of the functions,
ensuring real implementations are tested.

+++ lib/c/src/sysctl_helpers.c
@@ -5,7 +5,7 @@
 #include <string.h>
 #include <stdint.h>
 
-extern int64_t _syscall6(int, int, int, int, int, int, int);
+extern int64_t _syscall6(int, intptr_t, intptr_t, intptr_t, intptr_t, intptr_t, intptr_t);
 #define SYS_SYSCTL 243 // Matches sys/syscall.h syscall table
 
 int sysctl(int *name, unsigned int namelen, void *oldp, size_t *oldlenp, void *newp, size_t newlen) {
@@ -17,7 +17,7 @@ int sysctl(int *name, unsigned int namelen, void *oldp, size_t *oldlenp, void *n
         errno = EINVAL;
         return -1;
     }
-    int ret = (int)_syscall6(SYS_SYSCTL, (int)name, (int)namelen, (int)oldp, (int)oldlenp, (int)newp, (int)newlen);
+    int ret = (int)_syscall6(SYS_SYSCTL, (intptr_t)name, (intptr_t)namelen, (intptr_t)oldp, (intptr_t)oldlenp, (intptr_t)newp, (intptr_t)newlen);
     if (ret < 0) {
         errno = -ret;
         return -1;
